MODELS-1641 `src` is an MklDnnData object on the stack declared outside of the for loop scope. `src = MklDnnData<T>(&cpu_engine);` within the for loop calls the copy assignment constructor, which is essentially doing this: 1) Create a new MklDnnData object on the stack 2) Copy it into `src`
The problem here is the destructor never gets called by the copy
constructor, see:
https://en.wikipedia.org/wiki/Rule_of_three_(C%2B%2B_programming)

, and in particular, `src.user_memory_` never gets deallocated (Other members
could leak too--it just happens that user_memory_ is the issue in this case).
We need to keep user_memory_ around (`inputs` has a reference to it)
until issue we sum it here:
```
      std::vector<primitive> net;
      net.push_back(sum(sum_pd, inputs, dst.GetOpMem()));
      stream(stream::kind::eager).submit(net).wait();
```

So probably the best solution is to create multiple MklDnnData objects
outside of the for loop and let the stack destroy them all at the end of
the function which is what we do here.